Moved authors model inside the module directory
